Saving id instead of user
The cookie saves the id of the user as a string in the cookie. When the
user is needed it's requested from the database.  Minimized the use of
pointers to places where the value can be nil. Restructured parts of the
code to minimize the amount of knowledge each file has of the others.